requirements.txt
spotipy==2.24.0 
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
python-dateutil==2.9.0.post0n
thefuzz==0.22.1
rich==13.7.1